adds tracing for AArch64
Some parts of this code are shared with the 32-bit arm, specifically
tracing of flags and conditions.
For AArch64, we store operands only for the flags that are actually
used, as opposed to existing arm code, which always dumps all of them
unconditionally. To account for this change, loaded_cpsr and store_cpsr
have been refactored to now be bitfields indicating the individual
flags, rather than being just single booleans.
Old code just sets them to TRACE_CPSR_ALL and can be made more
fine-grained successively in the future if desired.